requirements.txt
Flask==2.0.1
Flask-RESTful==0.3.9
PyJWT==2.1.0
requests==2.26.0
pytest==6.2.4
